feat: created bullets component
