+++ react/components/ProductList/styles/product-list.css
@@ -20,6 +20,7 @@
   gap: 1rem;
   overflow-x: auto;
   scroll-behavior: smooth;
+  scrollbar-width: none;
 }
 
 .productListEmptyState {
